add room for differenc solvers
