+++ src/main.rs
@@ -1,21 +1,25 @@
 mod board;
 mod cell;
 mod examples;
+mod solver;
+
+use solver::NonogramSolver;
+
 extern crate z3;
 
 fn main() {
-    let mut board = examples::easy_board();
+    let board = examples::easy_board();
     board.print();
-    board.solve();
-    board.print();
+    let z3_nonogram_solver = solver::sat::Z3Solver::new(board);
+    z3_nonogram_solver.solve().expect("solver failed").print();
 
-    let mut board = examples::hard_board();
+    let board = examples::hard_board();
     board.print();
-    board.solve();
-    board.print();
+    let z3_nonogram_solver = solver::sat::Z3Solver::new(board);
+    z3_nonogram_solver.solve().expect("solver failed").print();
 
-    let mut board = examples::extreme_board();
+    let board = examples::extreme_board();
     board.print();
-    board.solve();
-    board.print();
+    let z3_nonogram_solver = solver::sat::Z3Solver::new(board);
+    z3_nonogram_solver.solve().expect("solver failed").print();
 }
