+++ src/solver/mod.rs
@@ -1,0 +1,7 @@
+use crate::board::Board;
+
+pub mod sat;
+
+pub trait NonogramSolver {
+    fn solve(&self) -> Option<Board>;
+}
